Rename os => operatingsystem
We are using operatingsystem as command name. Therefore the entities
functions must use it too.
