+++ gsa/src/web/store/entities/__tests__/operatingsystems.js
@@ -20,24 +20,10 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-import {createAll} from './utils/main';
+import {testAll} from '../utils/testing';
 
-const {
-  loadEntities,
-  loadEntity,
-  reducer,
-  selector,
-  entitiesActions,
-  entityActions,
-} = createAll('os');
+import * as funcs from '../operatingsystems';
 
-export {
-  loadEntities,
-  loadEntity,
-  reducer,
-  selector,
-  entitiesActions,
-  entityActions,
-};
+testAll('operatingsystem', funcs);
 
 // vim: set ts=2 sw=2 tw=80:
