+++ gsa/src/web/store/entities/reducers.js
@@ -35,7 +35,7 @@ import {reducer as group} from './groups';
 import {reducer as host} from './hosts';
 import {reducer as note} from './notes';
 import {reducer as nvt} from './nvts';
-import {reducer as os} from './os';
+import {reducer as operatingsystem} from './operatingsystems';
 import {reducer as ovaldef} from './ovaldefs';
 import {reducer as permission} from './permissions';
 import {reducer as portlist} from './portlists';
@@ -66,7 +66,7 @@ const entitiesReducer = combineReducers({
   host,
   note,
   nvt,
-  os,
+  operatingsystem,
   ovaldef,
   permission,
   portlist,
